prefect_topic_thunder/requirements.txt
attrs==19.3.0
Automat==0.8.0
blinker==1.4
boto3==1.16.33
botocore==1.19.33
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==20.3
cloudpickle==1.6.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
croniter==0.3.36
cryptography==2.8
Cython==0.29.21
dask==2020.12.0
dataclasses==0.6
dbus-python==1.2.16
distributed==2020.12.0
distro==1.4.0
distro-info===0.23ubuntu1
docker==4.4.0
ec2-hibinit-agent==1.0.0
entrypoints==0.3
fastapi==0.62.0
filelock==3.0.12
fire==0.3.1
future==0.18.2
hdbscan==0.8.26
HeapDict==1.0.1
hibagent==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
importlib-resources==3.3.0
incremental==16.10.1
Jinja2==2.10.1
jmespath==0.10.0
joblib==0.17.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
llvmlite==0.35.0
MarkupSafe==1.1.0
marshmallow==3.9.1
marshmallow-oneofschema==2.1.0
more-itertools==4.2.0
msgpack==1.0.1
mypy-extensions==0.4.3
natsort==7.1.0
netifaces==0.10.4
nltk==3.5
numba==0.52.0
numpy==1.19.4
oauthlib==3.1.0
packaging==20.7
pandas==1.1.0
pendulum==2.1.2
pexpect==4.6.0
plotly==4.14.1
prefect==0.13.19
protobuf==3.14.0
psutil==5.7.3
psycopg2-binary==2.8.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydantic==1.7.3
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==0.10.1
PyNaCl==1.3.0
pynndescent==0.5.1
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.3
python-box==5.2.0
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-dotenv==0.15.0
python-slugify==4.0.1
pytz==2020.4
pytzdata==2020.1
PyYAML==5.3.1
regex==2020.11.13
requests==2.22.0
requests-unixsocket==0.2.0
retrying==1.3.3
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
SecretStorage==2.3.1
sentence-transformers==0.3.9
sentencepiece==0.1.91
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sortedcontainers==2.3.0
sos==4.0
SQLAlchemy==1.2.14
ssh-import-id==5.10
starlette==0.13.6
systemd-python==234
tabulate==0.8.7
tblib==1.7.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==2.1.0
tokenizers==0.9.3
toml==0.10.2
toolz==0.11.1
torch==1.7.0
tornado==6.1
tqdm==4.54.1
transformers==3.5.1
Twisted==18.9.0
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ufw==0.36
umap==0.1.1
umap-learn==0.5.0
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
websocket-client==0.57.0
zict==2.0.0
zipp==1.0.0
zope.interface==4.7.1
